TESTS PASS except AvailCompilerTest#newAvailTest(), which fails due to name re-export not working.
Primitives now have a signature (see #blockTypeRestriction()).  Primitive *blocks* must inversely conform to that signature.  They can have more restrictive argument types, but must have a less (or equally) restrictive return type.

Various Avail code was shuffled (generally for bootstrapping purposes) to allow stronger types on primitives.

Added TupleDescriptor#from(...).

+++ Avail/src/com/avail/compiler/node/BlockNodeDescriptor.java
@@ -493,7 +493,7 @@ public class BlockNodeDescriptor extends ParseNodeDescriptor
 	 */
 	public static AvailObject newBlockNode (
 		final List<AvailObject> argumentsList,
-		final short primitive,
+		final int primitive,
 		final List<AvailObject> statementsList,
 		final AvailObject resultType)
 	{
